packages/market-pulse/skills/market-pulse-scraper/requirements.txt
yt-dlp
PyYAML==6.0.2
python-dateutil==2.9.0.post0
feedparser==6.0.11
requests==2.31.0
